feat: [#48 #45] 서식지 정보 fetch 로직 custom hook 으로 분리
서식지 정보 받아오는 로직 useHabitatInfo hook로 분리
hook 적용(FeedScrollBox,HabitatPreview)
전환 애니메이션 딜레이 시간 변경
